backend/requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
tensorflow>=2.8.0
scikit-learn>=1.0.0

# Alternative data processing
sentinelhub>=3.4.0
geopython>=0.1.0
sec-edgar-downloader>=2.0.0

# ML/AI libraries
gym>=0.21.0
stable-baselines3>=1.5.0
pymc3>=3.11.0
arviz>=0.11.0

# Quantitative finance
QuantLib>=1.24
scipy>=1.7.0

# Data streaming
confluent-kafka>=1.8.0
avro-python3>=1.10.0

# Cloud infrastructure
google-cloud-aiplatform>=1.12.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Development tools
pytest>=6.0.0
black>=21.5b2
flake8>=3.9.0


Flask>=2.0.0


bcrypt>=3.2.0


PyJWT>=2.0.0

# Additional required dependencies
SQLAlchemy>=1.4.0
ruptures>=1.1.0
dask>=2021.0.0
websockets>=10.0
dash>=2.0.0
torch>=1.10.0
pymc>=4.0.0
requests>=2.26.0

# Database connector
pymysql>=1.0.0
